ofp-actions: Make parse_reg_load() accept set_field syntax.
Commit 7eb4b1f1d "ofp-actions: Support OF1.5 (draft) masked Set-Field,
merge with reg_load." purportedly merged the set_field and reg_load
fields, including the commandline parsing to use the metafield syntax
like "load:192.168.0.1->OXM_OF_IPV4_SRC". However, the actual reg_load
parsing code was not updated.

This patch takes advantage of the newly refactored parse_set_field()
helpers to make parse_reg_load() conform to the documentation, which
allows more thorough validation of inputs, and allows reg_load to parse
fields that are wider than 64 bits, as well as masked fields.

Note that this disallows syntax that was previously allowed, such as:
    load:0x0a000001->OXM_OF_IPV4_SRC

Signed-off-by: Joe Stringer <joestringer@nicira.com>
---

Note that v2.4.0 includes the aforementioned commit; we should consider
either backporting this fix to branch-2.4, or reverting the
documentation modification on branch-2.4 to be consistent with the
previous syntax.